+++ Core/EditorCommand.cs
@@ -64,6 +64,8 @@ internal enum EditorCommand
     NextTab, PrevTab,
     GoTabNumber,
     ListTabs,
+    JumpBack,
+    JumpForward,
     SplitPane,
     NextPane, PrevPane,
     GoPaneNumber,